+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Fix instrumentation for very old versions of `redis` (0.12.1).
+
 ## 1.131.0
 - Limit the number of depdendencies collected as snapshot data.